requirements.txt
dj-database-url==1.2.0
psycopg2-binary==2.9.5
django==4.1.7
gunicorn==20.1.0
whitenoise==6.4.0
Brotli==1.0.9
openai==0.27.2
python_dotenv==1.0.0